Refactor into smaller components, implement editing of variables
