+++ src/api/vars/query/useVars.ts
@@ -1,6 +1,5 @@
-import { VarContentSchema, VarListSchema } from "../schema";
-
 import type { QueryFunctionContext } from "@tanstack/react-query";
+import { VarListSchema } from "../schema";
 import { apiFactory } from "../../utils";
 import { useApiKey } from "../../../util/store/apiKey";
 import { useNamespace } from "../../../util/store/namespace";
@@ -37,40 +36,6 @@ export const useVars = () => {
       apiKey: apiKey ?? undefined,
     }),
     queryFn: fetchVars,
-    enabled: !!namespace,
-  });
-};
-
-const getVarContent = apiFactory({
-  url: ({ namespace, name }: { namespace: string; name: string }) =>
-    `/api/namespaces/${namespace}/vars/${name}`,
-  method: "GET",
-  schema: VarContentSchema,
-});
-
-const fetchVarContent = async ({
-  queryKey: [{ namespace, apiKey, name }],
-}: QueryFunctionContext<ReturnType<(typeof varKeys)["varContent"]>>) =>
-  getVarContent({
-    apiKey: apiKey,
-    urlParams: { namespace, name },
-    payload: undefined,
-  });
-
-export const useVarContent = (name: string) => {
-  const apiKey = useApiKey();
-  const namespace = useNamespace();
-
-  if (!namespace) {
-    throw new Error("namespace is undefined");
-  }
-
-  return useQuery({
-    queryKey: varKeys.varContent(namespace, {
-      apiKey: apiKey ?? undefined,
-      name,
-    }),
-    queryFn: fetchVarContent,
     enabled: !!namespace,
   });
 };
